Split Monolithic Table into smaller components and fix measuring for lastColumnWidth
